autogen, peercount, logging updates
This change primarily is cleanup and logging minimizations. Part of this is avoiding endless
spam of the warning "No connected peers and no known addresses to dial!" when running a \
single node network. This lead to some autogen fixes.

This also makes any errors from config file parsing more helpful, suggesting that it is possibly
an old version of the config file, and only showing the first unrecognized field rather than
pages of errors, as is the case with an old config file.