node: start connect backoff and network.Notifiee fix
This makes the startup peer connection (handled by `maintainMinPeers`) use exponential
backoff on a per-peer basis, just like the reconnect logic started when a peer disconnects
(initiated by the `network.Notifiee`)

This also fixes the regression with the `network.Notifiee` being de-registered from the Host.

Finally, this has the `P2PService` register dummy stream handlers for all of the "required"
protocol IDs.  They are replaced with the actual handlers in `Node`  startup when the CE
is ready to take over.