Leader replacement support for Kwild
This PR introduces new ways to replace leader especially for scenarios where
the resolution based approaches such as consensus ParamUpdates is not possible
such as when leader is offline and not producing any blocks.

Validator node can express it's intent to replace the leader by issuing the
below command:
"kwild validators replace-leader <leaderCandidate> <height>" command.
This makes the node accept the block proposals from the new leaderCandidate
rather than from the previous leader. If majority of validators issue this
command to replace the leader with the new leaderCandidate at the specified
height, the leader change will be accepted.

Ensure that the new leaderCandidate should also issue this command, else it
will never propose the block.

If the old leader is still online and proposing blocks and the nodes started
the leader replacement process, the network can end up in the following states
depending on majority:
- If Majority of validators still agree on the previous leader, these
  validators will not accept the blocks from the new leaderCandidate and the
  blocks proposed by the leader will get the majority of the Acks thereby
  commiting the block from the previous leader. All the nodes that agreed on
  the leader change will revert back to accepting the old leader as the block
  committed has majority of the Validators votes. These nodes should retry the
  leader replacement process at future heights and ensure that they have
  majority votes for the replacement to be successful.
- If majority of the validators agreed to replacing the leader from a specific
  height, the block proposed by the new candidate would receive the majority
  votes and get committed to the chain. The previous leader and nodes
  supporting the previous leader will update their leader to the new one.
- If previous leader and the new leader both doesn't have the majority, the
  network is halted till one of them has majority of the votes.

BlockAnnounce messages are now processed even by the leaders as there can be
nodes assuming that they are the leaders because of the "replace-leader"
command, but the committed block is proposed by a different node, so these
nodes should be able to process this committed block and update the leader
param to the correct leader.