test: update stress tool for v0.10
This updates the `test/stress` tool for v0.10 Kwil DB.

You must have to have the key to `db_owner` to deploy a test schema namespace, and use `-key` to
specify (only secp256k1, and just the hex bytes, no key type).

When it makes a new namespace, take note of the printed namespace.  Then to run multiple instances,
copy that randomly generated namespace and use that with a second instance with `-ns` and a
different (or random key).

(previous stress tool just had each instance depoy a new db, but that's not really possible and more
unless each instance uses the same key, and that causes horrible nonce issues)

Say the DB owner is set to an ethereum address with a known corresponding private key.  Start the
stress tool with the `-key` flag using that key:

    ./stress -key 0610bb3b2c52c4c06e9ba7a2ce2a507f618c5ca7c93b7b3e7601a7f5182b7133

This will automatically create a new namespace in which it will start calling actions:

    [INF] STRESS: act schema: deployed namespace "stress_ca4yfwzg"

(I'll improve that log).

Grab that "blahblahxnvxnqod" string and use it with a second instance using the `-ns` flag, but either
a different key or no key (random):

    ./stress -ns blahblahuw4q1dd2