ce crash must stop node (#1366)
The ensures that fatal CE errors (those which kill the consensus event
loop other than parent context cancellation) are returned to the
the caller of Start. This fixes a halted / dead CE not stopping the
node.

This also reduces a lot of duplicate error logging when the error is
continually passed up the call stack.