rpc,config: disable RPC response compression by default, add setting
This was a hotspot in performance tests. This is not feature that needs to be enabled
by default. It can be a convenience to reduce response sizes if there are large
compressible responses AND the operator doesn't want to deploy a proper reverse
proxy to offload this task. Hence, this disables it by default and provides the option
to enable it for quick bandwidth reduction at the expense of increased CPU.