core/types,node/types: add cached tx hash to ktypes.Transaction, create node/types.Tx
This updates `core/types.Transaction` a.k.a. `ktypes.Transation` with an
unexported `cachedHash` field, which is only populated or referenced via
the new `HashCache` method, rather than modifying the `Hash` method to
do this automatically.  A few reasons are outlined in the docs, but they
include concurrency concerns, needless allocs in many cases, the tx
mutability gotchas particularly in client side applications.

This also removes the `node/types.NamedTx` which was used in a handful
of places to pair the hash with the `core/types.Transaction`, replacing
`NamedTx` with `node/types.Tx` for brevity.  This new types adds a
constructor `NewTx` that computes the has on construction.  The `Hash()`
method of this type shadows the embedded `types.Transaction`'s method to
return this stored hash.  This should be less likely to become an issue
since it is "internal" to the `node/...` packages, and the hash being
computed in the constructor avoids any concurrency concerns.  Plus it is
documented explicitly as an _immutable_ transaction type.

The `node/types.Tx` with it's construction-time hash storage is
primarily used in mempool and the CE. The `HashCache` method is used in
other places that deal with a `core/types.Block`, which is necessarily
defined in terms of a `core/types.Transaction`.