add DB notices to tx result logs, and error handling for tx results
This change adds NOTICE calls to the transaction execution / call result.

It also separates execution errors (which are created by calling ERROR() or by things like primary key conflicts) from internal errors (like the DB being OOM). When calling a view action, it returns this as a separate field. When executing an action, it includes the error as the final log.

Finally, it fixes an issue where ERROR could not be called in a SQL statement (which should have been supported).