docker: fix GOWORK build
The go.work file from the builder machine is not necessarily right
for the builder, which specified a Go version in the Dockerfile
such as 'FROM golang:1.23 AS build'.

Since we converted to the Dockerfile doing GOWORK builds, the solution
is to do 'go work init . ./core ./test' inside the builder.